Agent: add symf support and enabled keyword context by default (#2641)
Fixes https://github.com/sourcegraph/jetbrains/issues/256

Previously, symf was disabled in the agent because the download+unzip
logic was failing in Node.js. This PR fixes that problem (along with
    several other small issues), and adds a test case to stress the code
path where the user submits a chat message with enhanced context
enabled. This test case is disabled for now because symf sends an LLM
request that we can't record/replay (since symf is written in Go). The
test can be manually enabled locally to test the functionality. It's
still useful to merge this PR because it unblocks testing symf context
with JetBrains.

**Important**: this PR changes the default context to be keyword-based
instead of embeddings-based. This is aligned with our plans for February
GA. It's possible to change it back to be embeddings-based by adding the
custom configuration:

```json
{
  "cody.useContext": "embeddings"
}
```

Additionally (and optionally), use the following setting to disable symf
to get the old default before this PR:

```json
{
  "cody.experimental.symfContext": false
}
```


## Test plan


* Green CI
* (optional) Manually enable the enhanced context test case to see it
running locally. Even in replay mode, it still takes ~2 seconds to run
because it does an LLM request for query expansion inside symf. Also,
the assertion doesn't stress the context functionality because the
workspace folder is a tmp directory without git repo.
* Manually test with JetBrains in a real-world git repository and see if
the context is any more intelligent.
<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->