Symf: move term expansion into VS Code extension
Previously, we shelled out to `symf expand-query` to do term expansion
on the query. This approach prevented us from writing stable tests with
enhanced context because the LLM request inside the `symf` binary was
not recorded by Polly. This PR fixes the problem by reimplementing the
`expand-query` subcommand in TypeScript so that we can record the HTTP
request.