skip (instead of hanging indefinitely on) non-text context files
If a file is `@`-mentioned that is inaccessible or unreadable, Cody chat would hang indefinitely. Now, it skips the file, shows the user an error message, and proceeds with the chat response.

The most likely causes for a file being inaccessible or unreadable are: the file is binary, or the file was removed after the user selected it.

Fix https://github.com/sourcegraph/cody/issues/2390.