Symf: move term expansion into VS Code extension (#2644)
Previously, we shelled out to `symf expand-query` to do term expansion
on the query. This approach prevented us from writing stable tests with
enhanced context because the LLM request inside the `symf` binary was
not recorded by Polly. This PR fixes the problem by reimplementing the
`expand-query` subcommand in TypeScript so that we can record the HTTP
request.

## Test plan

See new integration tests for `expandQuery`. Ideally, we will tweak the
HTTP recording format to only return the latest completion event so PRs
like this don't have to be a 6k+ diff.

<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->