Agent: add new `chat/restore` endpoint
Previously, it was not possible to start a chat session with context
from another conversation. The lack of this feature prevented clients
like JetBrains from supporting preserving the chat history after
restarting the IDE. This PR fixes the problem by adding a new
`chat/restore` endpoint that allows the client to pass the transcript of
an old conversation.

This PR also makes the logic related to creating new chats more
robust. Previously, it relied on racing promises aginst hardcoded timeouts
to capture webview instances. Now, the chat-related commands return the
webview directly avoiding the need for brittle logic inside the agent.