VSCode: Remove close button from Enhanced Context Settings popup (#2680)
#2628 will use &times; in the enhanced context settings popup for
removing a repository, so we need to remove the close button. This
removes the close button. Now, the popup is closed by:

- Clicking outside the popup.
- Pressing Escape.

In addition, this takes care that keyboard focus goes to the
enable/disable checkbox when the popup opens, and returns to the
enhanced context button when the popup closes.

## Test plan

```
pnpm test:e2e -- context-settings.test.ts
```

Manual test:

1. Sign in and start a chat.
2. Open the enhanced context popup.
3. Press Space a few times, this should toggle the checkbox.
4. Press Escape, this should close the popup.
5. Press Space again, this should re-open the popup.
6. Click somewhere else on the chat. This should close the popup.