Improve Knip config (#2726)
After being _very glad_ to see Knip being used in this repository, I saw
the configuration and the slightly uncommon blend of the `vscode` and
`agent` workspaces (`"../agent/**/*.ts"`), I couldn't resist to check
out the repo and come up with some suggestions, as presented in this PR.

Looks like most of the results are (still) correct and there's even more
to clean up, yay! I'm happy to leave the actual clean up to developers
more familiar with the codebase (I'm also unaware of potential external
factors/consumers).