Agent: make `pnpm update-agent-recordings` more stable and run faster
Previously, the `update-agent-recordings` command did an
`rm -rf agent/recordings`. This was problematic because 1) it was slow
to update all the recordings to record one new HTTP request and 2) it
made the tests fail more regularly because the LLM response is random
even with temperature=0. This PR addresses both problems by allowing
Polly to manage deletion of unused recordings instead of doing
`rm -rf agent/recordings`. This means that updating recordings is very
fast because it only sends network requests for missing recordings, and
it also makes the tests a lot more stable because we keep old recordings
around for much longer.

One risk with this change is that we never update recordings. Polly
suports configuring an expiry date for recorings but we can't use that
at the moment because we hardcode the request date to minimize diffs. We
should consider removing that hardcoded date because it's no longer
needed after this PR (because we keep recordings around for much longer).