Symf: move term expansion to separate file (#2732)
Previously, I was unable to run the symf tests locally because it
reported an issue with `vscode` imports. This error was understandable
because the `symf.ts` file imported from the `'vscode'` module. Moving
the term expansion function to a separate file fixes the problem.
However, I'm not sure how the tests passed in the first place unless
something regressed recently.


## Test plan

Green CI
<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->