encapsulate @sourcegraph/cody-shared, restrict subpath imports
You should only import from `@sourcegraph/cody-shared` and not `@sourcegraph/cody-shared/{dist,src}/...`. This is to (1) enforce a clean boundary between the internal and external API of the shared lib, (2) avoid accidentally bundling multiple copies of the shared lib.

We may still remove the shared lib. In that case, this PR will still be a cleanup and footgun-avoider in the interim, and it will help us create a simpler API for clients even in a possible monolib.