remove old recipes code (#2746)
Recipes were replaced by commands in November 2023 (version 0.18.0).
Most of this code was unreachable (including after
https://github.com/sourcegraph/cody/pull/2719 was merged to make the
agent use commands not recipes). The ChatQuestion was still used for e2e
testing, so it remains, but with all recipe remnants stripped out.



## Test plan

CI, and manually test that chat and commands still work in VS Code.