Add "Sourcegraph Code Search" to TESTING.md (#294)
AFAIK `Code Search` features is not a priority for the upcoming GA
release, but I decided to briefly describe those functionalities &
expectations. The motivation is the latest error reported by David +
during testing it turned out that there is a problem with one of these
actions (at least on my IDE).