Add "Open GitHub Issue" error submitter (#296)
Unplanned, but I've made decision to add it, especially since
implementing it was a small effort. Convenient issue reporting will help
us receive more exceptions, so we'll be able to react to them by Feb GA
and increase the discoverability of the project. Maybe someone will be
engaged enough to participate in our open source project, who knows?

Related to: #135

## Test plan

1. Add `throw RuntimeException()` somewhere in the plugin code to
simulate exception.
2. Run the plugin and trigger that exception. An error notification will
appear.
3. Open notification, then click `Open GitHub Issue`.
4. Browser will open at the `issues/new` with useful context (plugin
version, IDE version, stacktrace, etc). You should be able to submit
that issue (but you shouldn't do it if the exception is only for testing
purposes!).