remove support for codebase-specific ignores (#2774)
Workspace ignores (`.cody/ignore` files at the root of a workspace) are
still supported. Previously, we also stored the mapping of a workspace
ignore file to the codebases it applied to, but we never look up by
codebase anymore (after https://github.com/sourcegraph/cody/pull/2692),
so the mapping can be removed. There is no loss or change in
functionality.

(I missed deleting this code in that PR.)