remove Transcript and most of createClient
Previously the primary consumer of these was the e2e test. OldChatQuestion was only used in the transcript test. There are some transcript tests that may still be useful that were deleted, but they were heavily reliant on an old and decaying implementation (OldChatQuestion), so they were low value. Transcript is still used by the inline edit workflow.