remove e2e, e2e-inspector, Transcript, most of createClient (#2778)
Previously the primary consumer of these was the e2e test.
OldChatQuestion was only used in the transcript test. There are some
transcript tests that may still be useful that were deleted, but they
were heavily reliant on an old and decaying implementation
(OldChatQuestion), so they were low value. Transcript is still used by
the inline edit workflow.

Discussion:
https://sourcegraph.slack.com/archives/C05AGQYD528/p1705482895338749