Autocomplete: do not adjust insert text in a VS Code-specific way for the agent (#2573)
Extracts the VS Code specific `insertText` processing 
(which makes all completions start at the beginning of the line no
matter what the current cursor position is) into a separate function
and it's now called separately. If we decide to change this behavior,
for the agent, it will be easy to do in the future.