Autocomplete: Make sure to not crash when the site config was changed (#2783)
This is an attempt to fix a potential rough experience for our
customers:

In the scenario where a server admin changes the completion model from
the default to e.g. `fireworks/starcoder`, we don't have a way for the
clients to be notified about this change. I think it's reasonable to
require clients to reload until those changes take effect.

The issue though, is that previously in the Anthropic code path, we
never included a `model` identifier in the requests. This means that
when the server receives the code completion request, it will fall back
to the _default_ that is configured - This has changed, though. That
will mean the Anthropic prompt will be sent to the new Fireworks
provider which is obviously not going to work (alone for the fact that
it's not a chat model).

This is one attempt at fixing this by passing through the model that is
defined in the upstream url. The catch? The allow list was updated only
a few months ago:
https://github.com/sourcegraph/sourcegraph/commit/70b068c4c6d90e88cdc023df964b846986dacb7b
which means that for Sourcegraph servers before this allow list, passing
a `model` might be coming back with errors instead.

There are two potential ways of fixing this:

1. Instead of fixing it on the client, we can make the server look at
the payload. If the prompt looks like the anthropic prompt and no model
overwrite is defined, route it to anthropic instead of Fireworks.
2. Handle the case of a model-not-found error on the client and, if
throw, back off from this and pass `undefined` through again.

This PR implements number 2) since avoids cases where admins never set
up Anthropic and yet for some reason the prompt is set to Anthropic.

## Test plan

We now sent the `model` param with Fireworks request. Screenshot is from
S2 where this works flawlessly. If the admin would change the site
config, we would still be able to make requests.

<img width="1810" alt="Screenshot 2024-01-17 at 16 07 16"
src="https://github.com/sourcegraph/cody/assets/458591/4da810bd-b942-4adf-9482-02521526eea6">

If the server does not have an allow list, we back of from this behavior
after the first error (we don't replay on error since code completions
are ephemeral anyways and if you don't get one, it's not the end of the
world):

<img width="1810" alt="Screenshot 2024-01-17 at 16 07 16"
src="https://github.com/sourcegraph/cody/assets/458591/885f218b-d27f-4763-855a-c490ca2d9892">