Agent: always disable local embeddings (#2796)
Previously, local embeddings were enabled for agent clients. We had
disabled local embeddings in the tests through an environment variable
because they were causing stability issues, but clients like the
JetBrains plugin were not setting this environment variable. This PR
fixes the problem so that local embeddings are always disabled for all
agent clients.


## Test plan

Our tests are already running in this environment.
<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->