resolve remote embeddings fileNames to URIs
This fixes an issue with remote embeddings where attempting to run a chat that fetched remote embeddings would completely fail with an error from trying to dereference `.scheme` on an undefined value.

The Sourcegraph GraphQL API returns remote embeddings results with `fileName` fields. We need to resolve this to a `uri` so that we don't pass around a relative fileName that we don't know how to resolve. To resolve it, we assume (as do other context sources) that the context is for the first workspace folder in VS Code. We need to pass this around as `workspaceFolderUri`, which is ugly but is the smallest change that will make this work. The remote embeddings code is going away soon anyway, so IMO this solution suffices.

Co-authored-by: Quinn Slack <quinn@slack.org>