Agent: update squirrel test to assert context files
Previously, the squirrel test didn't assert that the context files were
empty. Also, the squirrel test explicitly opened the squirrel.ts file so
it wasn't stressing the code path where symf context returns results
from unopened files. This PR fixes both problems, we remove the explicit
`textDocument/didOpen` notification for squirrel.ts and we add an
assertion that squirrel.ts is part of the context files in the
transcript.