Autocomplete: Cancel requests that are no longer relevant (#2855)
This PR adds the ability to abort inflight requests when the document
was updated in a way that yields new completions no longer relevant.

This is useful as an optimization to avoid streaming completions run for
too long, for example in cases where the user updates the document after
the request was started in a way that the completion would no longer
yield useful results (e.g. they rewrite the prefix). Or in case of
hot-streaks when the completion starts to diverge from the current
document prefix (e.g. when the user rejects the first completion and
thus all hot-streaks being generated would be useless).

The exact cases of when something is deemed relevant are tested and
document separately. Most importantly we leave a room for some typos by
the user to avoid cancelling requests as soon as the user makes one
wrong keystroke (even though they might revert it right away).

## Test plan

- See the added tests
- In real world, you can see this working by enabling the hot-streak
feature and observing the console for request cancellation
notifications.

---------

Co-authored-by: Valery Bugakov <skymk1@gmail.com>