Agent: add support for inline edit command "Document code" (#2870)
Previously, the agent only supported commands that appeared in the chat
view (explain, smell and test). This PR adds support for a new "Document
code" command that can automatically insert a docstring into the file.
This functionality requires expanding the agent protocol to allow the
server to update the document contents:

- `commands/document`: request sent from client to server to trigger
"Document code" command. This command returns a "task ID", which we can
use to track progress on this command.
- `editTaskState/didChange`: notification sent from the server to the
client to notify when the task enters a new phase (applying, formatting,
applied, finished, errored, ...).
- `textDocument/edit`: request sent from the server to the client to
apply a diff against the file contents
- `codeLenses/display`: notification sent from server to client to
display code lenses (buttons to accept/retry/undo generation of
docstring)
- `textDocument/didSave`: request sent from the client to the server to
notify the user has saved the file. Cody uses this as an implicit signal
that the docstring was accepted by the user. Users can also explicitly
accept via code lenses.

Additionally, this PR adds two now properties
`ClientInfo.capabilities.{codeLenses,edit}` for the `initialize`
handshake to allow the client to opt-into supporting the edit/codeLens
requests.

Finally, to get the functionality working as expected, this PR
implements folding ranges with a basic algorithm that follows
indentation.

## Test plan

See updated tests.
<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->