Expand context panel tree on EDT (#345)
This is a fix for [the recent exception I'm seeing in
nightly.](https://github.com/sourcegraph/jetbrains/issues/344)

I unfortunately was not able to repro with a local build. So I'm not
sure how it's getting called from a non-EDT thread -- some sort of race
condition, possibly. But this looks like the right fix -- I tried it
with/without, and the new context panel expands the tree properly for me
in both cases.

## Test plan

Manual testing. Passed.

---------

Co-authored-by: Steve Yegge <stevey@sourcegraph.com>