Run embeddings and symf retrieval in parallel and implement basic fusion (#2804)
This PR is changing the simple chat context engine to run the symf
search together with any dense retrieval strategies (local/remote
embedding) in parallel and fuse the results. The fusion is currently
done naively by allocating up to 80% of the context window for the symf
results and the remaining 20% for embeddings.

We will follow this up with a more advanced fusion logic that will also
address some issues for Autocomplete. We, however, need to adjust the
logic we have for RRF to have a finer granularity before we can do this.

## Test plan

- Added unit test for the fusion code
- Ask questions and look at the context items being picked up - Most
importantly it passes the squirrel test.