Agent: add ability to disable network requests to sourcegraph.com (#2909)
Previously, the agent sent log events to sourcegraph.com for enterprise
accounts even if any of the following things were true:

- `cody.serverEndpoint` was set to the enterprise URL.
- `cody.telemetry.level` was set to `off`. This setting is, in fact,
entirely ignored by the agent.

The root problem was that the v1 event logger was still logging one
event to sourcegraph.com without any option to route this request only
to the connected instance. This was problematic because some enterprise
instances are airgapped and this request ended up timing out.

This PR fixes the problem by adding support to exclusively send logging
events to the connected instance. This behavior can only be enabled by
setting the environment variable `CODY_LOG_EVENT_MODE` to the value
`'connected-instance-only'`. Previously, it defaulted to the value
`'dotcom-only'` or `'all'` without any workaround. We are using an
environment variable because that is enough to unblock JetBrains. It
probably makes sense to add a similar setting for the VSC extension.


Paired with @chwarwick on this PR
## Test plan

See updated enterprise test case that adds an assertion that we don't
send any requests to other URLs than the connected instance.

<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->