Agent: fix bug causing recording files to not remove unused files
Previously, the HTTP recording file kept growing bigger and bigger even
for PRs where we didn't add new requests and we had enabled the setting
`keepUnusedRequests: false`, which should remove entries for unused
requsts. Turns out that we were parsing the
`KEEP_UNUSED_RECORDINGS=false` environment variable as the string
value `"false"`, which is truthy. This PR fixes the problem so that we
now correctly remove unused recordings.