IntelliJ can now debug the Agent (#432)
I wanted a decent debugging setup, so I modified the agent to be able to
be debugged by IntelliJ. This lets me set breakpoints on both sides of
the protocol.

This was a fairly straightforward change, but I have only used it for a
few hours, so there may be issues. However it is very useful to be able
to debug the agent, so I am committing it now.

There are two modes of operation supported: Cody can spawn the agent or
IntelliJ can spawn it. The latter allows Cody to talk to any remote
agent.

I added detailed instructions on how to set this up in CONTRIBUTING.md.

## Test plan

Locally/manually tested. This is a dev-only code path, and tricky to
unit test.

Co-authored-by: Steve Yegge <stevey@sourcegraph.com>