Autocomplete: fix actual position calculation for hot-streak completions (#2906)
- As a follow-up to https://github.com/sourcegraph/cody/pull/2881, I
started adding more tests. I bumped into the issue that I had created
with the last-minute fix in the PR where I removed the `actualPosition`
from `docContext` and switched to computing it manually on the spot. The
math was incorrect in some cases: we need to know the length of the
string at the target line number to calculate the correct position. This
PR reverts back to having this field to avoid the complexity of handling
edge cases where the injected completion overwrites `currentLineSuffix`.
- It also adds a helper function `acceptFirstCompletionAndPressEnter`,
to the return value of `getInlineCompletionsWithInlinedChunks`, making
hot-streak tests even more compact and easier to read.