Autocomplete: More Jaccard improvements and per-line RRF (#2898)
This PR makes some changes to the new jaccard similarity retriever and
implements per-line RRF:

- Jaccard similarity search will now no longer yield results that
overlap in the lines. This was happening previously since we were
creating results by sliding a window over the source document line by
line and ranking it individually. That meant that lines with high
overlap would result in multiple results and overlapping windows. We
then had logic to specifically merge overlapping ranges. In practice
though, often time this would run unbound and would end up having the
whole document in one range after all ranges were merged. This then lead
to only one context item being added that contains all of a file.
   
We fix this by adding another linear pass through the ranked windows and
doing some bookkeeping to know which lines were not part of any yielded
result yet. Since this is an optimization problem and we only do one
linear scan, this can end up not yielding results for every line in the
document (i.e if the window size is 5 lines and we have the best
matching snippet starting at line 3, we will never be returning a result
for line 1 and 2). This, however, is not really relevant in reality as
long as we contain the few highest ranked results per file.
   
Merging was also removed to avoid combining lesser important results
into a more important one and thus inflating the rank.

- The existing RRF implementation had a similar issue. Since we were
doing the RRF on the individual documents, we ended up almost
exclusively only picking snippets from a single file and adding a lot of
low-ranking snippets into the mix. To fix this, this PR adds support for
adding more than one ranking identifier per result. This means that we
can rank on _individual lines_ instead of the whole file. When combining
results, we would start with the highest ranked line across all
documents and would mix in all snippets with a range overlapping that
line. If there is a lower ranked snippet on a different line, it would
not be picked.

I assume that both changes together would result in a much more healthy
distribution of context snippets for Autocomplete so I’m thinking about
rerunning maybe the local-mixed context fusion experiment. Additionally,
I think this might fix some of the issues we're seeing with the new
jaccard similarity (that right now is not stat. sig better or worse than
the old implementation)

## Test plan

- Added test cases
- Play around with our different context strategies and make sure
nothing crashes