Edit: Remove usage of codebase context (#2950)
## Description

Remove codebase context as:
- We now support user provided context, which will generally be much
more accurate for the user
- This context is very rarely useful and adds latency to every request.
Related code is unlikely to be useful for edits. It's possible we could
use codebase context for code related to the instructions, but this is
unlikely to work well given most instructions are short. We should
revisit a mechanism to manually trigger codebase context.

## Test plan

1. Run edits
2. Check they output as normally


<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->