add: display cody icon in chat panel title (#2937)
This commit updates the package.json for the VS Code extension by
removing the condition that required the resourceScheme to be file for
the cody.menu.commands to be visible in the editor title. This change
allows the command to be visible regardless of the resource scheme, as
long as cody.activated is true, config.cody.editorTitleCommandIcon is
set, and the editor is not in readonly mode. This enhances the command's
visibility and usability in various contexts within VS Code, and shows
up in the chat panels.

## Test plan

<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->


### Before

cody icon doesn't show up in chat panel


![image](https://github.com/sourcegraph/cody/assets/68532117/c10012db-71bb-4f58-a815-88d6e1c4f564)


### After


![image](https://github.com/sourcegraph/cody/assets/68532117/9fac8b12-1020-41ad-9162-7e0cc7e19141)