Agent: disable reusing webview panels
Previously, the ID from `chat/new` was reused for `commands/document`
if no message was sent to `chat/new`. This behavior was due to a desire
to reuse webview panels in VS Code, which is not behavior that we want
to mirror for the agent. Now, we always create a new ID for these
different commands even if no message has been sent to the original
`chat/new` ID.