Autocomplete: Add experimental fast path mode for Fireworks on PLG (#2927)
A new completion mode that skips the Sourcegraph instance and directly
connects to Cody Gateway. It's enabled only for PLG and Fireworks (which
make up the majority of our completion request).

This turns out to be a complete implementation of [RFC
888](https://docs.google.com/document/d/1xCSHHuz5M4PX6MQLun_KOz8RfoZyf-wWXRuTCBi07PE/edit)
since it includes a fix for the two main functunalities that needs to be
added:

- The client now handles chunks that do not have a repetition of all
previously received tokens. Another small improvement in latency as we
reduce the network overhead
- I made sure that the rate limiting behavior is kept as it was before.
The only header that was missing was the one to infer wether a user can
upgrade. We can also get this from the current `authStatus`

It also supports streaming exclusively. That is okay, because: We
previously supported streaming only in Node.js envs with a server
version that actually responds with a node stream. We can gate for the
Node version (but it's unclear if the web version is even used at all)
and we know the server is recent since we only run this on PLG.

You can turn it on via feature flag or VS Code config:
`"cody.autocomplete.experimental.fastPath": true`

## Test plan

- Turn on `"cody.autocomplete.experimental.fastPath": true`
- Trigger single and multi-line completions and notice that cody gateway
is the upstream that is being logged

To test the rate limiting, go into the dotcom admin console and set your
quota to 1. Then:

### As a free use
<img width="872" alt="Screenshot 2024-01-30 at 17 58 57"
src="https://github.com/sourcegraph/cody/assets/458591/2fd9ceca-ce86-4101-9ecc-c9d78729ab5c">

### As a cody pro user

<img width="730" alt="Screenshot 2024-01-30 at 18 00 51"
src="https://github.com/sourcegraph/cody/assets/458591/ff84bd73-98d3-4375-830a-929c62bb5e5b">

---------

Co-authored-by: Rafał Gajdulewicz <rafax@users.noreply.github.com>