fix issue where natural language search panel disappears instead of showing results (#2981)
It was not initializing its environment with the VS Code workspace root
folders to generate accurate workspace-relative paths. I forgot to add
that to the search webview entrypoint (only added it to the chat webview
entrypoint).

Confirmed working:


![image](https://github.com/sourcegraph/cody/assets/1976/64b79d46-8b85-49c7-a4b4-1bdb3eb70d36)


## Test plan

Run a natural language search and confirm you see results instead of the
webview disappearing.