Agent: handle document URIs with exclamation marks
Previously, the agent threw an error for `textDocument/*`
notificattions when the URI included an exclamation mark. The error was
thrown to preserve an important variant that we can always parse and
render a URI back into the original string
```ts
vscode.Uri.parse(uri).toString() === uri
```
This variant was getting violated when the JB plugin sent document URIs
that included an exclamation mark in them, for example:

```
file:///Users/com.jetbrains/ideaIC-2022.1-sources.jar!/com/intellij/RequiresBackgroundThread.java
```

The `!` got encoded as `%21`. This PR "fixes" the problem by mutating
the `uri: string` property to get the `Uri.toString()` formatted URI.
t's not a super satisfying solution, it would be more desirable if we
can get `vscode.Uri.parse(..).toString()` to be a safe roundtrip.
However, this change gets the most important job done: we preserve the
invariant.