feat: test (edit) command & add tests code lenses (#2959)
CLOSE https://github.com/sourcegraph/cody/issues/1475 &
https://github.com/sourcegraph/cody/issues/1761

PR to introduce `/unit` command that is a smarter version of the `/test`
command.

### /unit command

- `/unit` command generates unit tests in chat (old `/test`)
- `/test` command generates unit tests in line
- if a test file exists, the new unit test suit will be appended to the
end of the current test file
- if a test file does not exist, cody will suggest a file name and we
will create a temporary test file with the new test suit appended along
with the import statements (if any)


![unit-test-file](https://github.com/sourcegraph/cody/assets/68532117/0bac781c-ee69-4e30-9697-9b891944a7b7)

### Add Tests code lenses

`add-tests` is a command that can be executed via the Cody Command
CodeLenses (when enabled) in test files. It is not available via the
command menu or context menu.

Behind internal unstable flag. 


![unit-test-case](https://github.com/sourcegraph/cody/assets/68532117/4eef46d6-7165-4cf5-8ad6-b0e555263052)

## Summary

This PR includes the following change:
- Refactor the /unit command, rename it to /test, and move it out of
unstable
- Rename the old /test command to /unit
- Add test-case command
- Add `executeUnitCaseCommand` to execute unit test cases
- Added support for specifying a `destination file` in the
executeEditArguments interface in the edit/execute.ts file.
- Added support for specifying a `destination file` in the executeEdit
method in the edit/manager.ts file.
- Removed the `NewFixupFileMap` import and added support for specifying
a `destination file` in the EditProvider class in the edit/provider.ts
file.
- Removed the `NewFixupFileMap` import in the FixupController class in
the non-stop/FixupController.ts file.

## Test plan

<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->

Added e2e test.

### Test

Run the `Test` command to confirm the newly generated tests are added to
a new / exisiting test file automatically for you.


![image](https://github.com/sourcegraph/cody/assets/68532117/a04a61b7-7a6b-485c-af76-d584f379b959)

### Add Tests

Enable Code Lenses in your Cody Setting


![image](https://github.com/sourcegraph/cody/assets/68532117/79a4ceaa-5f56-4c3a-9a33-0770d3057a74)

Locate the `Add Tests` code lenses in a test file to check if new test
cases have been inserted in line for you


![image](https://github.com/sourcegraph/cody/assets/68532117/c0a09081-a36e-4167-9ef0-e65fef60d0b7)


## Follow-up

- [ ] option to append to current selection instead of insert
- [ ] context improvement for test

---------

Co-authored-by: Tom Ross <tom@umpox.com>